Remove ICL metrics (#3243)
* minify configs

* filter test

* move add_vision_dataset_transform

* v1 remove ICL

* checkdown

* remove test

* lint

* clean up docstring
