--- composer/datasets/__init__.py
@@ -1,22 +1,0 @@
-# Copyright 2022 MosaicML Composer authors
-# SPDX-License-Identifier: Apache-2.0
-
-"""Natively supported datasets."""
-
-from composer.datasets.in_context_learning_evaluation import (
-    InContextLearningCodeEvalDataset,
-    InContextLearningDataset,
-    InContextLearningLMTaskDataset,
-    InContextLearningMultipleChoiceTaskDataset,
-    InContextLearningQATaskDataset,
-    InContextLearningSchemaTaskDataset,
-)
-
-__all__ = [
-    'InContextLearningDataset',
-    'InContextLearningQATaskDataset',
-    'InContextLearningLMTaskDataset',
-    'InContextLearningCodeEvalDataset',
-    'InContextLearningMultipleChoiceTaskDataset',
-    'InContextLearningSchemaTaskDataset',
-]
