+++ composer/utils/__init__.py
@@ -49,6 +49,7 @@ from composer.utils.import_helpers import MissingConditionalImportError, import_
 from composer.utils.inference import ExportFormat, Transform, export_for_inference, export_with_logger, quantize_dynamic
 from composer.utils.iter_helpers import IteratorFileStream, ensure_tuple, map_collection
 from composer.utils.misc import (
+    add_vision_dataset_transform,
     create_interval_scheduler,
     get_free_tcp_port,
     is_model_deepspeed,
@@ -136,6 +137,7 @@ __all__ = [
     'LocalEvalClient',
     'MosaicMLLambdaEvalClient',
     'partial_format',
+    'add_vision_dataset_transform',
     'VersionedDeprecationWarning',
     'is_compressed_pt',
     'CliCompressor',
