+++ composer/utils/misc.py
@@ -3,13 +3,17 @@
 
 """Miscellaneous Helpers."""
 
+import logging
 import math
 import socket
+import textwrap
 from contextlib import contextmanager
 from typing import TYPE_CHECKING, Callable, Optional, Set, Type, Union
 
 import torch
 from torch.nn.parallel import DistributedDataParallel
+from torchvision import transforms
+from torchvision.datasets import VisionDataset
 
 if TYPE_CHECKING:
     from composer.core import Event, State, Time
@@ -22,7 +26,10 @@ __all__ = [
     'get_free_tcp_port',
     'model_eval_mode',
     'create_interval_scheduler',
+    'add_vision_dataset_transform',
 ]
+
+log = logging.getLogger(__name__)
 
 
 def create_interval_scheduler(
@@ -234,3 +241,45 @@ def partial_format(s, *args, **kwargs) -> str:
             kwargs[key] = '{' + key + '}'
 
     raise RuntimeError(f'Failed to format string {s} after {max_iters} iterations.')
+
+
+def add_vision_dataset_transform(dataset: VisionDataset, transform: Callable, is_tensor_transform: bool = False):
+    """Add a transform to a dataset's collection of transforms.
+
+    Args:
+        dataset (VisionDataset): A torchvision dataset.
+        transform (Callable): Function to be added to the dataset's collection of transforms.
+        is_tensor_transform (bool): Whether ``transform`` acts on data of the type
+            :class:`~torch.Tensor`. The transform will be inserted before or after
+            :class:`~torchvision.transforms.ToTensor` depending on if this is ``True`` or ``False``
+            respectively. If :class:`~torchvision.transforms.ToTensor` is not present, the
+            transform will be appended to the end of the collection of transforms. (default: ``False``)
+
+    Returns:
+        None: The ``dataset`` is modified in-place.
+    """
+    transform_added_logstring = textwrap.dedent(
+        f"""Transform {transform} added to dataset.
+        Dataset now has the following transforms: {dataset.transform}""",
+    )
+
+    if dataset.transform is None:
+        dataset.transform = transform
+        log.warning(transform_added_logstring)
+    elif isinstance(dataset.transform, transforms.Compose):
+        insertion_index = len(dataset.transform.transforms)
+        for i, t in enumerate(dataset.transform.transforms):
+            if isinstance(t, transforms.ToTensor):
+                insertion_index = i
+                break
+        if is_tensor_transform:
+            insertion_index += 1
+        dataset.transform.transforms.insert(insertion_index, transform)
+        log.warning(transform_added_logstring)
+    else:  # transform is some other basic transform, join using Compose
+        if isinstance(dataset.transform, transforms.ToTensor) and not is_tensor_transform:
+            dataset.transform = transforms.Compose([transform, dataset.transform])
+            log.warning(transform_added_logstring)
+        else:
+            dataset.transform = transforms.Compose([dataset.transform, transform])
+            log.warning(transform_added_logstring)
