--- tests/datasets/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2022 MosaicML Composer authors
-# SPDX-License-Identifier: Apache-2.0
